FredMDQD

Documentation for FredMDQD.

FredMDQD.FredMD

FredMDQD.FredQD

FredMDQD.fred_transform

FredMDQD.fred_transform

FredMDQD.fred_transform

FredMDQD.fred_transform

FredMDQD.fred_transform

FredMDQD.fred_transform

FredMDQD.fred_transform

FredMDQD.fred_transform

FredMDQD.info

FredMDQD.search_appendix

FredMDQD.FredMD — Type

FredMD(path::String)
FredMD(d::Date)
FredMD()

Load Fred MD data.

Arguments

path::String: Path to a manually downloaded version of Fred MD.

d::Date: Date of a vintage.

Details

If no arguments are provided, the most recent version of Fred MD will be downloaded.

If a d::Date is provided, the Fred MD vintage corresponding to the date will be downloaded.

If a path::String is provided, then the Fred MD file at the path will be loaded.

Notes

Fred MD data is compiled by Michael W. McCracken at the Federal Reserve Bank of St. Louis. For more information check out the official website at https://research.stlouisfed.org/econ/mccracken/fred-databases/

FredMDQD.FredQD — Type

FredQD(path::String)
FredQD(d::Date)
FredQD()

Load Fred QD data.

Arguments

path::String: Path to a manually downloaded version of Fred QD.

d::Date: Date of a vintage.

Details

If no arguments are provided, the most recent version of Fred QD will be downloaded.

If a d::Date is provided, the Fred QD vintage corresponding to the date will be downloaded.

If a path::String is provided, then the Fred QD file at the path will be loaded.

Notes

Fred QD data is compiled by Michael W. McCracken at the Federal Reserve Bank of St. Louis. For more information check out the official website at https://research.stlouisfed.org/econ/mccracken/fred-databases/

FredMDQD.fred_transform — Method

Apply a FRED tranformation code

Arguments

tcode: FRED transformation code

x: a vector of data

Transformation Codes

The following transformation codes are currently supported

$x$

$\Delta x$

$\Delta^2 x$

$log(x)$

$\Delta log(x)$

$\Delta^2 log(x)$

``Δ(x{t+1} / xt - 1)

FredMDQD.fred_transform — Method

return x

FredMDQD.fred_transform — Method

return Δx

FredMDQD.fred_transform — Method

return Δ²x

FredMDQD.fred_transform — Method

return log(x)

FredMDQD.fred_transform — Method

return Δlog(x)

FredMDQD.fred_transform — Method

return Δ²log(x)

FredMDQD.fred_transform — Method

return Δ(xₜ₊₁ / xₜ - 1)

FredMDQD.info — Method

Provide information about FredMDQD.

FredMDQD.search_appendix — Method

search_appendix(s::Symbol, args..., kwargs...)

Search the appendix for variable names.

Arguments

s::Symbol: Should be :MD for Fred MD and :QD for Fred QD.

needle::Union{String, Regex}: Pattern to find in the appendix.

Keyword Arguments

case_sensitive::Bool=false: Should the search be case sensitive?

historic::Bool=false: Should the search be in the current or historic appendix?

Returns

Returns a DataFrame of all the matches in the appendix.